backend/requirements.txt
numpy>=1.21
opencv-python>=4.5

matplotlib>=3.5
scipy>=1.7
pandas>=1.3
easyocr>=1.6
onnxruntime>=1.15
mediapipe>=0.9
httpx>=0.23
fastapi>=0.89
uvicorn>=0.22
python-dotenv>=0.21
pydantic>=1.10
supabase==2.0.2
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
gunicorn==21.2.0
psycopg2==2.9.9